requirements.txt
Flask==2.2.2
opencv-python==4.7.0.72
numpy==1.24.2
Pillow==9.5.0
